change Navigation to appNavigation and move to navigation folder
