+++ app/containers/InAppNotification/index.tsx
@@ -5,7 +5,7 @@ import { dequal } from 'dequal';
 
 import NotifierComponent, { INotifierComponent } from './NotifierComponent';
 import EventEmitter from '../../utils/events';
-import Navigation from '../../lib/Navigation';
+import Navigation from '../../lib/navigation/appNavigation';
 import { getActiveRoute } from '../../utils/navigation';
 import { IApplicationState } from '../../definitions';
 import { IRoom } from '../../reducers/room';
