+++ app/containers/markdown/Table.tsx
@@ -3,7 +3,7 @@ import { ScrollView, Text, TouchableOpacity, View, ViewStyle } from 'react-nativ
 
 import { CELL_WIDTH } from './TableCell';
 import styles from './styles';
-import Navigation from '../../lib/Navigation';
+import Navigation from '../../lib/navigation/appNavigation';
 import I18n from '../../i18n';
 import { themes } from '../../constants/colors';
