+++ app/lib/methods/actions.ts
@@ -3,7 +3,7 @@ import { TRocketChat } from '../../definitions/IRocketChat';
 import EventEmitter from '../../utils/events';
 import fetch from '../../utils/fetch';
 import random from '../../utils/random';
-import Navigation from '../Navigation';
+import Navigation from '../navigation/appNavigation';
 import sdk from '../rocketchat/services/sdk';
 
 const triggersId = new Map();
