+++ app/sagas/encryption.js
@@ -4,7 +4,7 @@ import { put, select, takeLatest } from 'redux-saga/effects';
 import { ENCRYPTION } from '../actions/actionsTypes';
 import { encryptionSet } from '../actions/encryption';
 import { Encryption } from '../lib/encryption';
-import Navigation from '../lib/Navigation';
+import Navigation from '../lib/navigation/appNavigation';
 import { E2E_BANNER_TYPE, E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../lib/constants';
 import database from '../lib/database';
 import RocketChat from '../lib/rocketchat';
