+++ app/sagas/messages.js
@@ -1,7 +1,7 @@
 import { select, takeLatest } from 'redux-saga/effects';
 import { Q } from '@nozbe/watermelondb';
 
-import Navigation from '../lib/Navigation';
+import Navigation from '../lib/navigation/appNavigation';
 import { MESSAGES } from '../actions/actionsTypes';
 import RocketChat from '../lib/rocketchat';
 import database from '../lib/database';
