+++ app/views/NewMessageView.tsx
@@ -16,7 +16,7 @@ import { IApplicationState, IBaseScreen, ISearch, TSubscriptionModel } from '../
 import I18n from '../i18n';
 import database from '../lib/database';
 import { CustomIcon } from '../lib/Icons';
-import Navigation from '../lib/Navigation';
+import Navigation from '../lib/navigation/appNavigation';
 import RocketChat from '../lib/rocketchat';
 import { compareServerVersion } from '../lib/utils';
 import UserItem from '../presentation/UserItem';
